+++ Servidor/src/main/java/com/example/demo/Repository/IngredienteRepository.java
@@ -2,10 +2,9 @@ package com.example.demo.Repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-
 import com.example.demo.Entity.IngredientesEntity;
 
 @Repository
-public interface IngredienteRepository  extends JpaRepository<IngredientesEntity, Integer>{
+public interface IngredienteRepository  extends JpaRepository<IngredientesEntity, Long>{
 
 }
