+++ Servidor/src/main/java/com/example/demo/Repository/ProductoRepository.java
@@ -2,10 +2,9 @@ package com.example.demo.Repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-
 import com.example.demo.Entity.ProductoEntity;
 
 @Repository
-public interface ProductoRepository extends JpaRepository<ProductoEntity, Integer>{
+public interface ProductoRepository extends JpaRepository<ProductoEntity, Long>{
 
 }
